% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_creation_functions.R
\name{check_polygons}
\alias{check_polygons}
\title{Find the polygon in which a spatial point is located.}
\usage{
check_polygons(polygon_df, points_df, key, .progress = "text")
}
\arguments{
\item{polygon_df}{polygon data frame in which the spatial points are supposed to be located.}

\item{points_df}{data frame of points, consisting at least of a pair of lat/lon variables
as well as a unique ID column.}

\item{key}{character string naming the column of the spatial polygon data frame which uniquely identifies
each subregion.}

\item{.progress}{name of the progress bar (from plyr) to use.}
}
\value{
A data frame of unique point IDs and the respective ID for the subregion they are located in.
}
\description{
Given a spatial polygon data frame and a data frame of point coordinates,
this function finds the subregion each point is located in, assigns its
respective ID and returns a data frame
}

