% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_creation_functions.R
\name{extract_from_geocode}
\alias{extract_from_geocode}
\title{Extract coordinates and address from a geocode locations list.}
\usage{
extract_from_geocode(x)
}
\arguments{
\item{x}{list of geocoded locations generated by
\code{ggmap::geocode(locations, output = "all", source = "google", nameType = "long")}}
}
\value{
A data frame consisting of \code{address}, \code{lat} and \code{lon} of all locations.
}
\description{
\code{extract_from_geocode} extracts the exact address and coordinates from a list
   of locations generated with \code{\link[ggmap]{geocode}}.
}
\examples{
\dontrun{
extract_from_geocode(ggmap::geocode(c("Paris, France", "London, UK", "Berlin, Germany"),
output = "all", source = "google", nameType = "long"))
}
}

