% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_creation_functions.R
\name{read_chronicle}
\alias{read_chronicle}
\title{Scrape Events of Anti-Refugee Violence from "Mut gegen rechte Gewalt" Chronicle.}
\usage{
read_chronicle(years)
}
\arguments{
\item{years}{numeric vector of years for which to retrieve data from the chronicle.}
}
\value{
A data frame of events as listed on the website, consisting of columns for date, location, bundesland,
   category, summary and source.
}
\description{
This function is built only for one specific purpose:
Scraping the website \url{https://www.mut-gegen-rechte-gewalt.de/service/chronik-vorfaelle}
for all events a specified year (currently, 2015 and 2016). It retrieves the date, location,
bundesland, category, summary and source for each event and returns a data frame.
}
\examples{
\dontrun{
chronicle <- read_chronicle(c(2015, 2016))
}
}

