% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_creation_functions.R
\name{read_chronicle_year}
\alias{read_chronicle_year}
\title{Scrape Events of Anti-Refugee Violence in a Given Year.}
\usage{
read_chronicle_year(year)
}
\arguments{
\item{year}{numeric value specifying the year for which to retrieve data from the chronicle.
Currently, it only works with 2015 or 2016.}
}
\value{
A data frame of events from a single year as listed in the chronicle,
   consisting of columns for date, location, bundesland, category, summary and source.
}
\description{
This function scrapes all events in a specified year from the "Mut gegen rechte Gewalt" chronicle.
It retrieves the date, location,
bundesland, category, summary and source for each event and returns a data frame.
}

