% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_creation_functions.R
\name{read_data}
\alias{read_data}
\title{Scrape Events of Right-Wing Violence in 2015 from Online Source.}
\usage{
read_data(page)
}
\arguments{
\item{page}{integer between 0 and n (which is the current maximum number of subpages in the list of events minus 1).}
}
\value{
A data frame of events as listed on the website, consisting of columns for date, location, bundesland,
   category, summary and source.
}
\description{
This function is built only for one specific purpose:
Scraping the website \url{https://www.mut-gegen-rechte-gewalt.de/service/chronik-vorfaelle}
for all events in 2015. It retrieves the date, location, bundesland, category, summary and source
for each event and returns a data frame.
}

