% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_events.R
\name{split_events}
\alias{split_events}
\title{Split events to obtain unique categories.}
\usage{
split_events(df = NULL)
}
\arguments{
\item{df}{A data frame. If left unspecified, the \code{arvig} data
frame is used. A user-specified data frame must contain the columns
category_en and category_de.}
}
\value{
A data frame.
}
\description{
\code{split_events} splits multi-category events from \code{arvig} into
observations with unique categories.
}
\examples{

\dontrun{
data(arvig)
head(split_events())

arvig \%>\%
  filter(stringr::str_detect(category_en, "&")) \%>\%
  split_events()
}
}
